Move BaseChain and LLMChain to legacy entrypoint in core
