+++ langchain-core/src/legacy/chains/index.ts
@@ -1,0 +1,2 @@
+export * from "./base.js";
+export * from "./llm.js";
