+++ langchain-core/src/output_parsers/index.ts
@@ -6,3 +6,4 @@ export * from "./structured.js";
 export * from "./transform.js";
 export * from "./json.js";
 export * from "./xml.js";
+export * from "./noop.js";
